% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC-MS functions_TDU.R
\name{collate_rt_files}
\alias{collate_rt_files}
\title{Collate RT Files}
\usage{
collate_rt_files(
  alkane_files_dir,
  sample_info_file,
  output_dir = "Z:/courtney/"
)
}
\arguments{
\item{alkane_files_dir}{Directory containing alkane RT files}

\item{sample_info_file}{Path to sample info CSV file}

\item{output_dir}{Directory where results will be saved}
}
\value{
List containing processed reference data and updated sample info
}
\description{
Collects and processes alkane standard RT files for GC-MS analysis
}
