% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Creat_short_cut}
\alias{Creat_short_cut}
\title{Creat a shortcut for Massomics}
\usage{
Creat_short_cut()
}
\value{
None
}
\description{
This function will creats a handful shortcut to Massomics. By clicking the shortcut, user will get the enviorment setup and MassOmics GUI launched automatically.
}
\examples{
Creat_short_cut()

}
