% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_library_functions.R
\name{create_sub_library}
\alias{create_sub_library}
\title{create_sub_library}
\usage{
create_sub_library(
  workdir = tk_choose.dir(caption = "Select working directory")
)
}
\arguments{
\item{workdir}{locate the working Dir}

\item{output}{specify the output file name initial}

\item{Ion.bin}{the mz tolerance (Da) for peak integration}

\item{intensity_type}{the type of signal to be reported \code{"Peak Height"} or \code{"Peak Area"}}

\item{GGReport}{the reporting mode \code{"Slow"} plot every integrated peak, or \code{"Fast"} generate an overall report}
}
\value{
None
}
\description{
This is a function for peak integration according to the IDs in identification summary
}
\examples{
GCMS_integration()

}
