% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC-MS functions.R
\name{GCMS_integration}
\alias{GCMS_integration}
\title{GCMS_integration}
\usage{
GCMS_integration(
  output = "GC-MS Result",
  Ion.bin = 0.5,
  workdir = getwd(),
  intensity_type = c("Peak Area", "Peak Height"),
  GGReport = c("Fast", "Slow")
)
}
\arguments{
\item{output}{specify the output file name initial}

\item{Ion.bin}{the mz tolerance (Da) for peak integration}

\item{workdir}{locate the working Dir}

\item{intensity_type}{the type of signal to be reported \code{"Peak Height"} or \code{"Peak Area"}}

\item{GGReport}{the reporting mode \code{"Slow"} plot every integrated peak, or \code{"Fast"} generate an overall report}
}
\value{
None
}
\description{
This is a function for peak integration according to the IDs in identification summary
}
\examples{
GCMS_integration()

}
