% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC-MS functions_TDU.R
\name{gcms_rt_correction_workflow}
\alias{gcms_rt_correction_workflow}
\title{Complete GC-MS RT Correction Workflow}
\usage{
gcms_rt_correction_workflow(
  alkane_files_dir,
  sample_info_file,
  output_dir = "Z:/courtney/",
  derivatization = c("TMS", "MCF")
)
}
\arguments{
\item{alkane_files_dir}{Directory containing alkane RT files}

\item{sample_info_file}{Path to the sample info CSV file}

\item{output_dir}{Directory where results will be saved}

\item{derivatization}{Vector of derivatization types to process (e.g., c("TMS", "MCF"))}
}
\value{
List containing all processing results
}
\description{
Combines the collation of RT files and RT correction in one workflow
}
