% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassOmics.r
\name{MassOmics.GUI}
\alias{MassOmics.GUI}
\title{MassOmics.GUI}
\usage{
MassOmics.GUI()
}
\value{
None
}
\description{
This function initiate a graphical user interface for MassOmics. To initiate the GG-MS data processing software, open the R console and type the codes in the example to initate the user interface.
}
\examples{
MassOmics.GUI()

}
