% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassOmics.r
\docType{package}
\name{MassOmics}
\alias{MassOmics}
\alias{addKeggCodes}
\alias{MassOmics-package}
\title{\tabular{ll}{
Package: \tab MassOmics \cr
License: \tab GPL (>= 2)\cr
URL: \tab 
}}
\usage{
addKeggCodes(inputData, keggCodes, folder, save = TRUE,
  output = "Results with kegg codes", addCodes = TRUE)
}
\description{
\tabular{ll}{
Package: \tab MassOmics \cr
License: \tab GPL (>= 2)\cr
URL: \tab 
}
}
\author{
George GUO \email{George.GUO@auckland.ac.nz}
}
\keyword{package}
