% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_library_functions.R
\name{msp_to_msl}
\alias{msp_to_msl}
\title{msp_to_msl}
\usage{
msp_to_msl(mspfile = NULL)
}
\arguments{
\item{mspfile}{specify the mspfile file name with its location}
}
\value{
MSL library file Path
}
\description{
This is a function will convert the MSP library to MSL library. The default export folder will be the same as of source msp file.
}
\examples{
msp_to_msl()

}
