% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC-MS functions.R
\name{MZrange_autodetect}
\alias{MZrange_autodetect}
\title{mz range detection}
\usage{
MZrange_autodetect()
}
\arguments{
\item{QCfiles}{specify the data file(s) to find the mz range}
}
\value{
a vector contains mz lower/upper limit
}
\description{
This is a function that help user to determine the sharing mz range acrossing data files.
}
\examples{
MZrange_autodetect()

}
