% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Omics_htest}
\alias{Omics_htest}
\title{Students T-test and the analysis of variance (ANOVA) on Omics data sets}
\usage{
Omics_htest()
}
\value{
None
}
\description{
MassOmics R-package also includes standard statistical analyses such as the Students T-test and ANOVA, coupled with false discovery rate (FDR), Bonferroni correction, and Tukey HSD test to adjust p-values when conducting multiple comparisons. To open the user-interface for statistical analysis, click <Statistics> and <Run ANVOA & T-test>. The implementation of Students T-test or ANOVA is determined automatically by number of input conditions. After selecting the type of statistical corrections, p value filtering, and log transformation, click <submit> and a pop-out window will appear to allow the user to classify samples into each group. The final statistical result is saved in the chosen working directory.
}
\examples{
Omics_htest()

}
