% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Omics_PCA}
\alias{Omics_PCA}
\title{Partial least squares-discriminant analysis (PLS-DA) analysis on Omics data sets}
\usage{
Omics_PCA()
}
\value{
None
}
\description{
Partial least squares-discriminant analysis (PLS-DA) is used to compress the dimension of multivariate data, and thereby display outliers and relationships between samples or classification of conditions. Under <Statistics> click <Run PLSDA> and a user interface window will appear to adjust a number of conditions, sphere radius, text size, log transfor-mation, and model validations. To plot 3 dimensional PLSDA, click <submit> and a pop-out window will allow the user to classify sam-ples into each group. All parameters can be changed on the user inter-face of PLSDA and replotted instantly via the <Re-Plot> function.
}
\examples{
Omics_PCA()

}
