% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Omics_PLSDA}
\alias{Omics_PLSDA}
\title{Principal component analysis (PCA) analysis on Omics data sets}
\usage{
Omics_PLSDA()
}
\value{
None
}
\description{
Principal component analysis (PCA) is used to compress the dimension of multivariate data, and thereby display outliers and relationships between samples or classification of conditions. Under <Statistics> click <Run PCA> and a user interface window will appear to adjust a number of conditions, sphere radius, text size, log transfor-mation, and model validations. To plot 3 dimensional PCA, click <submit> and a pop-out window will allow the user to classify sam-ples into each group. All parameters can be changed on the user inter-face of PCA and replotted instantly via the <Re-Plot> function.
}
\examples{
Omics_PLSDA()

}
