% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC-MS functions.R
\name{PeakDiagnosis}
\alias{PeakDiagnosis}
\title{Peak diagnosis by overlay multiple chromatograms}
\usage{
PeakDiagnosis(path = "GC-Peak Diagnosis Report.csv", font.size = 0.5)
}
\arguments{
\item{font.size}{Size of report font}
}
\value{
None
}
\description{
The user could visually confirm all peaks, and determine whether the peak is due to background noise by selecting <Overlay All Chromatograms> and clicking the <RUN> button in the box of step 3. A popout window will show a figure that plots chromatograms for all identified metabolites. You can also adjust the font size of metabolite names if the display is inappropriate. You must disable the <Fast Mode> to plot overlay of all chromatograms.
}
\examples{
PeakDiagnosis()

}
