% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC-MS functions_TDU.R
\name{Rt_correction_model}
\alias{Rt_correction_model}
\title{RT Correction Function}
\usage{
Rt_correction_model(
  derivatization = "TMS",
  output_dir = "Z:/courtney/",
  peaks_list = NULL
)
}
\arguments{
\item{derivatization}{Type of derivatization ("TMS" or "MCF")}

\item{output_dir}{Directory where results will be saved}
}
\value{
List containing the corrected data and model information
}
\description{
Performs retention time correction on GC-MS data based on derivatization type
}
