% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomforest.R
\name{SERRF}
\alias{SERRF}
\title{Systematic error removal using random forest (SERRF) normalization}
\usage{
SERRF(
  input = "Area.csv",
  Predict_level = "QC",
  data = NULL,
  datatype = c("MSDIAL", "MASSOMICS"),
  infopath = NULL,
  Log_trans = F,
  zero_imputaion = T
)
}
\arguments{
\item{input}{locate the molecule intensities data}

\item{Predict_level}{specify the sample class that will used for cluster prediction}
}
\value{
None
}
\description{
This systematic error removal method using random forest algorithm to detect molecule cluster based on sequential intensities profile and use median intensity of this cluster to normalize the data and eliminat the unwanted systematic variations in large sample sets.
}
\examples{
SERRF(input = "Area.csv",Predict_level=c("QC","Sample"))

}
